requirements.txt
defusedxml==0.4.1
Django==1.7.4
django-allauth==0.19.1
django-filter==0.9.2
djangorestframework==3.0.5
Markdown==2.5.2
oauthlib==0.7.2
Pillow==2.7.0
python3-openid==3.0.5
requests==2.5.1
requests-oauthlib==0.4.2
selenium==2.44.0
